+++ src/solvers/sgd.jl
@@ -47,7 +47,7 @@ function solve(sgd::SGD, net::Net)
   end
 
   destroy_coffee_breaks(sgd, net)
-  map(destroy, param_history)
+  map(x -> map(destroy, x), param_history)
 end
 
 function update_parameters(net::Net{CPUBackend}, solver::SGD, learning_rate, state, param_blob, hist_blob, gradient, data_type)