+++ NEWS.md
@@ -10,6 +10,7 @@
   * Power, Split, Element-wise layers
   * Local Response Normalization layer
   * Channel Pooling layer
+  * Dropout Layer
 * Documentation
   * Complete MNIST demo
   * Detailed document on each component